+++ app/src/main/java/com/futsch1/medtimer/preferences/WeekendModePreferencesFragment.kt
@@ -68,7 +68,7 @@ class WeekendModePreferencesFragment : PreferenceFragmentCompat() {
 
     private fun updateTimePicker(settings: UserPreferences) {
         val preference = preferenceScreen.findPreference<Preference?>(WEEKEND_TIME)
-        preference?.setSummary(timeFormatter.minutesToTimeString(settings.weekendTime.hour * 60 + settings.weekendTime.minute.toLong()))
+        preference?.setSummary(timeFormatter.minutesToTimeString(settings.weekendTime.hour * 60 + settings.weekendTime.minute))
     }
 
     private fun setupTimePicker() {
@@ -78,7 +78,7 @@ class WeekendModePreferencesFragment : PreferenceFragmentCompat() {
                 val weekendTime = currentSettings?.weekendTime ?: LocalTime.of(9, 0)
                 timePickerDialogFactory.create(weekendTime.hour, weekendTime.minute) { minutes: Int ->
                     preferencesDataSource.setWeekendTime(LocalTime.of(minutes / 60, minutes % 60))
-                    preference1!!.setSummary(timeFormatter.minutesToTimeString(minutes.toLong()))
+                    preference1!!.setSummary(timeFormatter.minutesToTimeString(minutes))
                     requestReschedule()
                 }.show(parentFragmentManager, TimePickerDialogFactory.DIALOG_TAG)
                 true
